+++ WearAndTear/Config/Server/DecayModifiersConfig.cs
@@ -14,28 +14,28 @@ namespace WearAndTear.Config.Server
         /// Multiplier on decay (damage) caused by wind
         /// </summary>
         [DefaultValue(1)]
-        [Range(0, float.PositiveInfinity)]
+        [Range(0d, double.PositiveInfinity)]
         public float Wind { get; set; } = 1f;
 
         /// <summary>
         /// Multiplier on decay (damage) caused by humidity (rainfall)
         /// </summary>
         [DefaultValue(1)]
-        [Range(0, float.PositiveInfinity)]
+        [Range(0d, double.PositiveInfinity)]
         public float Humidity { get; set; } = 1f;
 
         /// <summary>
         /// Multiplier on decay (damage) caused by the passing of time (Time decay is usually a secondary decay type)
         /// </summary>
         [DefaultValue(1)]
-        [Range(0, float.PositiveInfinity)]
+        [Range(0d, double.PositiveInfinity)]
         public float Time { get; set; } = 1f;
 
         /// <summary>
         /// Multiplier on decay of molds
         /// </summary>
         [DefaultValue(1)]
-        [Range(0, float.PositiveInfinity)]
+        [Range(0d, double.PositiveInfinity)]
         public float Mold { get; set; } = 1f;
     }
 }
